+++ src/lein_monolith/task/each.clj
@@ -34,7 +34,7 @@
   (merge
     target/selection-opts
     {:parallel 1
-     :parallel-unordered 1
+     :no-dep-order 0
      :endure 0
      :report 0
      :silent 0
@@ -53,8 +53,8 @@
   (concat
     (when-let [threads (:parallel opts)]
       [:parallel threads])
-    (when-let [threads (:parallel-unordered opts)]
-      [:parallel-unordered threads])
+    (when (:no-dep-order opts)
+      [:no-dep-order])
     (when (:endure opts)
       [:endure])
     (when (:report opts)
@@ -304,7 +304,7 @@
             (print (str task-output))
             (flush)))
         ;; Print convenience resume tip for user.
-        (when-not (or (:parallel opts) (:parallel-unordered opts) (:endure opts))
+        (when-not (or (:parallel opts) (:endure opts))
           (let [resume-args (into
                               ["lein" "monolith" "each"]
                               (map u/shell-escape)
@@ -395,21 +395,22 @@
   serially or concurrently."
   [ctx targets]
   (let [parallel (get-in ctx [:opts :parallel])
-        parallel-unordered (get-in ctx [:opts :parallel-unordered])]
+        no-dep-order (get-in ctx [:opts :no-dep-order])]
+    (prn {:parallel parallel :no-dep-order no-dep-order})
     (cond
+      (and parallel no-dep-order)
+      (run-parallel-unordered! ctx (Integer/parseInt parallel) targets)
+
       parallel
       (run-parallel! ctx (Integer/parseInt parallel) targets)
-
-      parallel-unordered
-      (run-parallel-unordered! ctx (Integer/parseInt parallel-unordered) targets)
 
       :else
       (run-linear! ctx targets))))
 
 
 (defn- run-all!
-  "Run all tasks, using the `:parallel`, `:parallel-unordered`, `:silent`, and
-  `:output` options to determine behavior."
+  "Run all tasks, using the `:parallel`, `:silent`, and `:output` options to
+  determine behavior."
   [ctx targets]
   (if (or (get-in ctx [:opts :silent])
           (get-in ctx [:opts :output]))